+++ minecraft/protocol/text.go
@@ -1,12 +1,12 @@
 package protocol
 
 const (
-	TextCategoryMessageOnly = uint8(iota)
+	TextCategoryMessageOnly = iota
 	TextCategoryAuthoredMessage
 	TextCategoryMessageWithParameters
 )
 
-var textCategories = map[uint8][]string{
+var textCategories = map[int][]string{
 	TextCategoryMessageOnly:           {"raw", "tip", "systemMessage", "textObjectWhisper", "textObjectAnnouncement", "textObject"},
 	TextCategoryAuthoredMessage:       {"chat", "whisper", "announcement"},
 	TextCategoryMessageWithParameters: {"translate", "popup", "jukeboxPopup"},
